requirements.txt
streamlit
langchain-groq
langchain_classic
langchain_experimental
langchain-community
langchain_core
langgraph
pandas
matplotlib
seaborn
tabulate
python-dotenv
scikit-learn